実行ファイルを1つにまとめた
